refactor(api): project_configs.py 분할 완료 — global_configs·monitoring 추출, _legacy 소멸
전역 설정 6개를 global_configs.py로, 모니터링·요약 7개를 monitoring.py로
이동하고 남은 _legacy.py(list_projects + DB 필터 헬퍼)를 core.py로 rename.

1,818줄 1파일 → 11모듈 최대 269줄. 60개 라우트 HTTP 표면 동일.

include 순서 확정: global_configs → monitoring → skills·agents·mcp·hooks·
commands·memories → rules. 전역이 맨 앞인 이유는 {project_id} 자리에 리터럴
global 이 들어가 /global/rules 계열 5개를 삼키기 때문이다.

핸들러가 빠져나가 고아가 된 섹션 배너 10개를 core.py에서 제거했다 —
추출이 만든 잔해라 정리 대상이다. 라우트·핸들러는 무변경.

R2b(split_audit) 통과: 정의 61개 전부 바이트 동일, 중복·유실 0건.
